Rename add-trip-modal.js to save-trip-modal.js.
